refactor: SettingMock을 FakeSetting으로 변경
