+++ app/src/test/java/woowacourse/movie/FakeSetting.kt
@@ -1,8 +1,8 @@
-package woowacourse.movie.view
+package woowacourse.movie
 
 import woowacourse.movie.system.Setting
 
-class SettingMock : Setting {
+class FakeSetting : Setting {
     private val settings: MutableMap<String, Boolean> = mutableMapOf()
     override fun getValue(key: String, default: Boolean): Boolean {
         return settings[key] ?: default
